Add statistical super class to mahalanobis
